franky fixes a whole bunch of warnings to compile in a more pedantic fashion also some compiler flags have been added to visual studio build so that we are stricter and safer

+++ lib/renderer/FrameBuffer.cpp
@@ -77,7 +77,7 @@ void checkFrameBufferStatus(GLenum stat)
 }
 
 FrameBuffer::FrameBuffer(const Texture & texture, const RenderBuffer & depthBuffer)
-: mFrameBufferId(-1)
+: mFrameBufferId((GLuint)-1)
 {
 	GCLAssert(texture.IsValid());
